+++ core.ts
@@ -25,6 +25,7 @@ import type { default as Scripts, ScriptOrFunction } from "./core/scripts.ts";
 
 import type { default as FS, Entry, Loader } from "./core/fs.ts";
 import type Logger from "./core/logger.ts";
+import type Searcher from "./core/searcher.ts";
 import type Writer from "./core/writer.ts";
 import type IncludesLoader from "./core/includes_loader.ts";
 import type DataLoader from "./core/data_loader.ts";
@@ -133,6 +134,7 @@ export type {
   Scopes,
   ScriptOrFunction,
   Scripts,
+  Searcher,
   Server,
   ServerEvent,
   ServerEventType,