+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/),
 but not always is possible (due the use of unstable features from Deno).
 Any BREAKING CHANGE between minor versions will be documented here in upper case.
 
+## [Unreleased]
+### Added
+- New `site.searcher` property with a instance of `Searcher` class.
+  It's used by plugins like `search`, `nav`, `sitemap` and `feed`.
+  (Previouly, each plugin had it's own instance).
+
 ## [1.17.3] - 2023-05-10
 ### Fixed
 - The `lume/` import is not correctly generated with `lume init`.
@@ -2229,6 +2235,7 @@ The first version.
 [#418]: https://github.com/lumeland/lume/issues/418
 [#419]: https://github.com/lumeland/lume/issues/419
 
+[Unreleased]: https://github.com/lumeland/lume/compare/v1.17.3...HEAD
 [1.17.3]: https://github.com/lumeland/lume/compare/v1.17.2...v1.17.3
 [1.17.2]: https://github.com/lumeland/lume/compare/v1.17.1...v1.17.2
 [1.17.1]: https://github.com/lumeland/lume/compare/v1.17.0...v1.17.1